[flow] Array.filter uses a one-sided type-guard
Summary:
Array.filter only uses the positive branch to refine the resulting array, so it can just accept a one-sided type guard function.

This is in preparation of one-sided type guard rollout. Libdefs are allowed to use "implies" even when the feature is not enabled by the global flag. Array.filter will behave exactly like before, since the two-sided version is a subtype of the one-sided one.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D58003606

fbshipit-source-id: 9d6b9cac14a1f83225427055de13bea3644e9714