[flow][tests] Repurpose react_create_element test as the test for a best-effort userland typing for `React.createElement`
Summary:
In this next diff, I'm going to make `React$CreateElement` fully opaque. At the same time, we will provide OSS users who are not ready to get rid of all `React.createElement` call with an alternative, which they can use to override the libdef.

This diff creates the alternative, tests it, and later we can link it in the changelog.

Changelog: [internal]

Reviewed By: pieterv

Differential Revision: D58031931

fbshipit-source-id: f44d1d37c74901c7b2cd2e0ca98108c70179590d