[flow] show instantiated constructor type on hover over "new" in new expressions
Summary:
When hovering over the "new" keyword in `new C` (and the spaces between "new" and "C") we will now be showing the instantiated type of the class constructor. So, in
```
class C<X> {
  constructor(x: X) {}
}

const c = new C(1);
//        ^
```
we will show `(x: number) => void`

Implementation-wise, we use a similar approach to D47180573 and record the instantiated type in `ConstructorT`. Instead of adding a special entry in the typed AST to store it, we introduce a mapping from `New` expression locations to types in the context.

Changelog: [fix] Flow will now show the type of the constructor of a class (instantiated if it is generic) when hovering over the "new" keyword in `new C(e)`.

Reviewed By: SamChou19815

Differential Revision: D58122866

fbshipit-source-id: 1d9ae302e6370973a9d358dcfd981c6839813a95