[flow] Default `react.disable_function_components_default_props` to true
Summary: Changelog: [errors] In v0.237.2, we shipped the flag `react.disable_function_components_default_props` that will make Flow ignore `defaultProps` on function components for React component typing purposes. This flag is now on by default. If this is too disruptive for your codebase, you can turn this off for now. Note that we do not intend to keep this flag forever, so you should try to turn this flag on as soon as possible.

Reviewed By: gkz

Differential Revision: D57922496

fbshipit-source-id: 515f7772bc1626b752325310dc302394e02697f5