[flow] do not consider frozen flag as literal type
Summary:
In subtyping kit we have a shortcut that lets us bypass invariant checking of "object literal" types.

This shortcut is intended to leverage the idea of a unique reference to an object, but is implemented in an unsound way that lets us write:
```
const obj = { f: 1 };
(obj as { -f: number }).f = 2;
obj.f as 1; // oops
```
[try-Flow](https://flow.org/try/#1N4Igxg9gdgZglgcxALlAIwIZoKYBsD6uEEAztvhgE6UYCe+JADpdhgCYowa5kA0I2KAFcAtiRQAXSkOz9sADwxgJ+NPTbYuQ3BMnTZA+Y2yU4IwRO4A6SFBIrGVDGM7c+h46fNRLuKxJIGWh8MeT0ZfhYlCStpHzNsFBAMIQkIEQwJODAQfiEyfBE4eWw2fDgofDBMsAALfAA3KjgsXGxxZC4eAw0G-GhcWn9aY3wWZldu-g1mbGqJUoBaCRHEzrcDEgBrbAk62kXhXFxJ923d-cPRHEpTgyEoMDaqZdW7vKgoOfaSKgOKpqmDA+d4gB5fMA-P6LCCMLLQbiLOoYCqgh6-GDYRYIXYLSgkRZkCR4jpddwPfJLZjpOBkO4AX34kA0SVs9gABBA0AArdkAXnZwHZMGQ7IAjOz6QBuAA6UDlAAoubyMCRBezFiL2cIRDdJQBKKwwfnsgBMsvlUGVRvZqvFUtyIAaJhIcGgSQaAAYrKaAMwAdh9IHpQA)

Until now "frozen" object follow this special behavior. As frozen objects we consider both the result of `Object.freeze`, but also the exported object of a module. The unsound behavior shown above is relevant in both cases. Consider, the multi-file example
```
==> export.js <==
declare var f: 1;
export default { f };

==> updater.js <==
import E from './export';
export function update() {
    (E as { f: number }).f = 2;
}

==> main.js <==
import E from './export';
import {update} from './updater';
update();
E as { f: 1 }; // no error today
```
This allows us to cast the export of export.js to `{f: 1}` after having assigned `2` to the `.f` property.

Changelog: [errors] We are making the typing of the exported object of a module more strict. If you see errors when casting the exported type of a module to a writable object, try either casting it to a readonly version instead, or casting the exported object to the writable object type.

Reviewed By: SamChou19815

Differential Revision: D57191363

fbshipit-source-id: 43094ee2b617cabc66bc4ebea9f45ba50786dbb0