[flow][refactor] Construct react jsx props type with children within statement.ml
Summary:
This is the diff that starts to take advantage of all the previous work and the hard work of killing `React.createElement` custom function. Without these baggages, what seems straightforward is finally possible: constructing the props object according the jsx evaluation rule, and pass the constructed props to `React.CreateElement` use_t, instead of doing it in the constraint system. It will allow us to finally starting to simplify the most complicated function that's impossible to type accurately in the userland.

Right now, the children prop is added later in object_kit after several layers of indirection:

https://github.com/facebook/flow/blob/9650ce00305f25a3ff3b77574109ff653053d5cf/src/typing/object_kit.ml#L295-L310

This might be a relic of the old constraint system, and the fact that the `React.CreateElement` use_t has to work for both the jsx form and the custom function call form.

This diff moves the addition logic inside `Statement.jsx_mk_props` when we are under React JSX. We also need to specialize the logic for fragment, by constructing an props object containing the single children prop.

With this cleanup, the `jsx_children` field in `React.CreateElement` is always `None`, which I will simplify in later diffs.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D58270807

fbshipit-source-id: b53883368bf902e8b4c462a76dd5e436297e5e27