[flow] do not concretize `try_singleton_throw_on_failure`
Summary:
The inputs to `try_singleton_throw_on_failure` in Speculation_kit are always resolved types, and therefore concretized, or tvars we are solving for and so should not be the target of concretization. This means that we can bypass this part of speculation and move directly to the part of triggering TryFlow. This part could be simplified a bit, but I've kept it for now since the rest of the handling of `SingletonCase` relies on speculation machinery.

This simplification allows us to remove the need to specify `upper_unresolved` when calling into `try_singleton_throw_on_failure`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D58392206

fbshipit-source-id: d86cf7c7655524b716e2c5655ca0365a7495b6ae