[flow] Introduce the concept of symbol
Summary:
In this diff, I will finally introduce the concept of symbol, as a replacement of reason where we actually care about the semantics of what's tracked with symbol.

The design is inspired by [TS' symbol](https://github.com/microsoft/TypeScript/blob/d8086f14b6b97c0df34a0cc2f56d4b5926a0c299/src/services/services.ts#L639-L653). It should have the following properties:

1. Tracks only definition instead of use sites.
2. Must be named
3. Must be trustworthy

Not all reasons have a path forward to become a symbol, and this is fine. Notably, while all the types must produce a reason, the same does not need to hold for symbols, because not all types are mapped to a named symbol.

In this diff, I only introduced the structure and nothing else. In the next diff, we will start to apply it to modules and namespaces.

Changelog: [internal]

Differential Revision: D58387128

fbshipit-source-id: 06fe1cf8dcd8d0d8934f5d1797e3a3c205313154