[flow][refactor] Track symbol of module and namespace in the types
Summary:
In this diff, we start to keep track of symbols. We start with tracking symbols in the namespace type. For now, it will be populated with both `SymbolNamespace` and `SymbolModule`, mostly to preserve the existing behavior that shows `module 'abc'` when we hover on `NS` in `import * as NS from 'abc'`.

In the next diff, we will going to use this information to reliably recover names of modules and namespaces in normalizer.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D58387130

fbshipit-source-id: f1b24792a1dd1c4ccdcea194a47c61c85e4fe6c1