[flow] Use symbol information to generate ty_symbol for modules and namespaces, unbreak autocomplete for namespaces
Summary:
In this diff, we start to take advantage of the symbol tracking for namespaces, so that we can reliably recover the namespace symbol information in ty_normalizer to create declarations.

Currently in ty_normalizer, when we see a `NamespaceT`, we check whether the reason is a reason that looks like a module, and if so, we create a `ModuleDecl`. It has two issues:

1. Unreliable reason inspection
2. It doesn't account for the fact that namespace declaration does exist...

Instead of deciding based on reasons, we can now do this based on symbols. For namespaces, we now create `ModuleDecl` or `NamespaceDecl` based on whether we see a `SymbolModule` or `SymbolNamespace`. We no longer need to consider all kinds of reasons that might end up being modules or namespaces, but instead rely on the more structured symbol information.

With this change, we actually fixed the autocomplete for namespace qualified types like `type Foo = React.ComponentType`, since we are now correctly generating a `NamespaceDecl`.

Changelog: [ide] We now provide autocomplete for types in the global `React` type-only namespace.

Reviewed By: panagosg7

Differential Revision: D58443649

fbshipit-source-id: b1590856e15fb80268f3d96d4996571aa16521d1