[flow][cleanup] Make ImportModuleNsTKit.on_ModuleT return a tuple of `(values_type, types_tmap)` instead of `NamespaceT`
Summary:
The job of this function is to compute the type information of a namespace, consisting of the value part and type part separately. The status quo of returning a NamespaceT is more a legacy of the dark past of not having `NamespaceT`  and using `ObjT` with `TypeT` fields to hack it.

As a result, this function no longer needs the namespace symbol. In places that want to construct the `NamespaceT`, we can do it separately.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D58478500

fbshipit-source-id: c6e5021fa4cef2d24de3caaa4e71d74b16868d9a