[flow][hover] improvement in union printing
Summary:
When printing unions of long types, the result would look something like the following
```

LongLongLongLongLongLongLongLongName
|
AnotherLongLongLongLongLongLongLongLongName
```
Notice the extra line in the beginning and that every `|` is on a line of its own.

This diff fixes this by making this pattern a bit more compact:
```
LongLongLongLongLongLongLongLongName
| AnotherLongLongLongLongLongLongLongLongName
```
Potential future work is to prepend a `|` to the first member when the union is multiline.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D58508654

fbshipit-source-id: d6fc0c234b5301e0bddd2f557ecdd40c3fe3efbd