[flow] Stop depending on repositioning cache for fully resolved tvars
Summary:
The repositioning loop doesn't happen on its own. It happens because the tvar being repositioned is cyclic itself.

For fully resolved tvars, we have builtin defense against the cyclic behavior since LTI 2.0, so it shouldn't happen, so we don't need to use the repositioning cache to defend it.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D58552974

fbshipit-source-id: 4babade5f5ee08d0d31e24a772b5a8ff7ef4bc07