[flow] Fork core to unblock OCaml 5
Summary:
It's discovered [during the attempt to make Flow build on all platforms with OCaml 5](https://github.com/facebook/flow/pull/9166) that, Core, the library we use, cannot be compiled with windows, due to the use of the non-platform-independent `endian.h`.

We can try to bump or wait, but since we only need a very small and specific part of core (Hash_queue used to power check_cache), it might be better to just fork and extract the part necessary.

In this diff, I did exactly that by forking the `Hash_queue` as `Core_hash_queue`, as well as a few modules it depends on. The fork doesn't contain everything, since we only use a small subset of APIs, I only kept a small set of APIs and deleted most of the APIs, especially some used collection APIs like map and filter that will end up pulling in more modules.

With this, we killed the final dependency on core, and we can remove it now.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D58768628

fbshipit-source-id: f8aa6364137e5adde04f152942343c509d4734bc