[flow][find-ref] Fix missing results when there is no explicit access
Summary:
We are missing jsx props result in find-ref, for very trivial reasons... We do the has_symbol check since D6849882, back when there is actually an optimization to avoid type checking. However, now we will require type artifacts for find-ref anyways, this is no longer useful.

The code has since drifted so that the has_symbol check no longer covers all cases where a prop can occur. For example, it's missing jsx props, which is exactly the reason for the missing results exposed in the previous diff.

Let's just remove the check to prevent further drifting that would be hard to catch, so that find-refs can reliably work.

Changelog: [fix] Fixed a bug that causes missing results in find-ref and rename

Reviewed By: alexmckenley

Differential Revision: D58969408

fbshipit-source-id: 628f69153decc26e54ace7d3912b82e3a77efbc1