[flow] do not abort refactoring on type synthesis error
Summary:
Before this diff in examples like
```
declare function bar(x: any): void;
function Foo(x: <A very large type>) {
  const z = bar(x);
  return null;
}
```
When selecting `const z = bar(x);` we would expect to get an option to extract to a separate function in global scope. We did not get this option, however, and this was due to a "type too large" error that we would get when trying to normalize <A very large type>.

This diff changes the behavior in situations like this by allowing the extraction/refactoring to proceed. Instead of adding the large normalized type, it just leaves the position empty.

So the above example would become:
```
declare function bar(x: any): void;

function Foo(x: <A very large type>) {
  newFunction(x);
  return null;
}
function newFunction(x): void {
  const z = bar(x);
}
```
Notice that the `x` parameter does not have an annotation instead of getting <A very large type>.

In terms of implementation, the main change is in making `synth_type` return a result type, instead of throwing an exception on error.

Changelog: [fix] Made the "extract to function" refactoring more robust to errors during code synthesis.

Reviewed By: SamChou19815

Differential Revision: D59099403

fbshipit-source-id: 1453149a3e129d80c3ed351c4eb4f6c389935e9e