[flow] Hardcode property names for arrays/maps/sets for better errors
Summary:
It's fundamentally confusing that calling `.pop` on an Array<T> produces an error message dealing with $ReadOnlyArrays. This is unfortunately very hard to fix in Flow's current reason system--the Array --> ReadOnlyArray transformation occurs as the result of a typeapp, and repositioning a TypeAppT doesn't correctly reposition its result :/

To get around this for the vastly common cases, we just hardcode the properties of Arrays that aren't present in ROArrays (ditto for sets and maps) and directly produce a useful error message when they're looked up, rather than swapping for an ROArray and letting the property lookup machinery produce the error.

Reviewed By: jbrown215

Differential Revision: D59075411

fbshipit-source-id: 3f7f9f0c93723a9762f5cadc86b8d90be992fb84