[flow] allow React immutable values to flow to explicitly deeply read only values
Summary: As an exception to the rule that React$Immutable types don't escape, if we can determine that the type that such a value flows into is itself deeply readonly, then we can admit the flow. This is likely not a perfect analysis, but it will help in cases where users are confused that a `React.Immutable<Array<number>>` is incompatible with `$ReadOnlyArray<number>`.

Reviewed By: jbrown215

Differential Revision: D59081021

fbshipit-source-id: 27de679ef3fb0beabeb26988b5a75105ca1691af