[flow] show symbols used in types at type-at-pos
Summary:
Changelog: [feature] hover will show a list of all the symbols found in the inferred type and the locations where they are defined.

Vscode LSP and CLI are supported. The LSP version includes a link to the definition. The CLI version only shows the name of the file (no path). flow_dot_js is not yet supported (no change there).

Reviewed By: SamChou19815

Differential Revision: D59138352

fbshipit-source-id: c8f6bc5adc0b126382766231063e38b465412bf9