[flow] compute set of symbols used in types at type-at-pos
Summary:
Adds a set of `Loc.t Ty.symbol`s to `type_at_pos_result`. Note that we're converting from `ALoc.t` to `Loc.t` to avoid dealing with the conversion in the printing step.

Will be used in the next diff as part of the hover response.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D59250838

fbshipit-source-id: cedbc046f6f78076f11e867914778d281db7ea85