[flow] Introduce a JSDoc purpose variant for get-def, where we jump to constructor definition for `new`
Summary:
The behavior change of D44585123 should only be scoped for jsdoc purposes. Fortunately, we can use the `purpose` param I introduced during find-ref work for this. Only under `purpose = JSDoc`, get-def on `new Identifier` will jump to the constructor of `Identifier`.

Changelog: [fix] Go-to-definition on `new C()` will jump to definition of `C` instead of the constructor definition. Hovering on `C` will still show the jsdoc on the constructor.

Reviewed By: panagosg7

Differential Revision: D59071365

fbshipit-source-id: 5c8719db4afac114fc378e0184396b5c5ff1afc1