[flow] React deep read only makes sets and maps readonly
Summary:
Currently maps and sets are not handled by React deep-read-onlyness, which allows React rule violations to easily be introduced. This diff alters that, by having class instances track whether they are ReactDRO or not. If they are, and their class ids correspond to maps or sets, we'll replace the type with a corresponding $ReadOnlyMap/Set when we flow the instance to an operation that could call `.set()` or `.add()`.

In addition, since *all* instances know whether they are DRO or not, we can now get *limited* enforcement of deep read onlyness on instances of other classes--we can prevent properties from being written to and propagate DROness on property reads. We cannot yet, though, prevent method calls that mutate the self.

Reviewed By: jbrown215

Differential Revision: D58962644

fbshipit-source-id: cc8d549ead92d83528f287de0ebbcdbfe9a2e342