[flow][refactor] Consolidate various asserts use_t into a single RunTypeAssertion
Summary:
The main recursive function of flow_js is getting to the point that it's so big that we are getting errors like

```
/var/folders/v3/_pdcw7dj3qvgvlm80f7rkqgc0000gn/T/build_096570_dune/camlasme35b1e-e9c742.s:19019:2: error: fixup value out of range
 tbnz x7, #0, L1174
```

on ARM Macs. What's going on? We are jumping between too many lines in assembly...

https://github.com/facebook/flow/commit/5422a344243a92e8795e728b3eac7b839f82731a has pushes it over the edge and https://github.com/facebook/flow/commit/427e0614186fdcd7f3e2fbd77bc979ddc2b83a34 brings it back over the edge, but it's clear that we are still dangerously close to the point of failing the build again.

This diff tries to give us a little bit of room by consolidating various type assertion use_t into a single one. It makes flow_js slightly shorter, and also helps to encode the idea that all these assertions are concretized at the single point.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59248767

fbshipit-source-id: 106fb100b8cac3ec1bddc0414e93d9ef4ec0d9e0