[flow] format hover type refs in command handler (#9184)
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/9184

This is a much better design. The server shouldn't need to know about the client specifics to adapt the location formatting. This should happen in the client like with most other commands.

This diff fixes achieves this by storing `(string * Loc.t)`s in the hover response in ServerProt, instead of raw strings.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D59309580

fbshipit-source-id: c2f5036dcea0a8e13f7bc9b1746035b7ceebeff9