[flow] Undo D42970432 to remove FullyResolved special case
Summary:
D42970432 tries to fix an action-at-a-distance behavior. However, I think it has less to do with FullyResolved vs Resolved, but more to do with use_op attached to FullyResolved and Resolved. Now that use_op attached to FullyResolved and Resolved are fully removed, we no longer need to special case.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59358030

fbshipit-source-id: 61d7e9abec09d276f690c19780e7b21f6676f9b1