[flow] Make the `$Values` utility type work for arrays and tuples as well
Summary:
Changelog: [feature] You can use the `$Values` utility type to get the type of the elements of an array type or tuple type.

We often tell people to use indexed access with `number`, e.g. `type T = Array<number>; T[number]`, but this is not an intuitive syntax to read or write. Instead, why not have `$Values` work for arrays and tuples as well?

Reviewed By: SamChou19815

Differential Revision: D59563339

fbshipit-source-id: 23ce2fe98398d97ed6fecef28557e7357a370725