[flow] Stronger defense against cyclic tvars
Summary:
This diff ensures that: for any local bad cyclic tvar, it will always be forced to the error type.

Why doesn't the current system enforce this desirable property? Well, let's consider what happens in this simple case if we just force every tvar without unwrapping:

```
type A = A;
```

In this case, we will have something like

```
id1 <-- TypeT (OpenT(id1))
starting state = {
   valid = unforced lazy (TypeT (OpenT(id1)))
   state = Unforced
}
```

When we force the state, the following will happen:

```
# right before forcing lazy
state = {
   valid = unforced lazy (TypeT (OpenT(id1)))
   state = Forcing
}

# forcing lazy, which trigger forcing id1 again
we are in the `Forcing -> on_error` case
nested forcing of id1 returns any, added error

# back to the outer forcing
# it doesn't know what's happening inside,
# so the outer one still forced to lazy (TypeT (OpenT(id1)))
state = {
   valid = forced lazy (TypeT (OpenT(id1)))
   state = Forced
}
```

Therefore, without unwrapping, all it does is to raise an error, without providing any downstream guarantees. Downstream uses like repositioning will still see something cyclic like this

```
t = ___ -->
     /\   |
     |    |
     <-----
```

so they still need to rely on other things to prevent non-termination (e.g. caches).

Unwrapping does help this case, since it makes the above example to evaluate to `TypeT (AnyT)`. However, due to the arbitrary depth limit and the things we decide to unwrap, it's not a 100% solution.

That's why, in this diff, we introduce a new state for fully resolved lazy tvars: `ForcedWithCyclicError`, which we will enter when we detected that the tvar is cyclic. When we finished forcing the state, if we find ourselves in this state, then we will stay in this state and use the error type forever, instead of using the lazy type forcing result.

With this, we will be able to use this to simplify the unwrapping by only unwrapping one level shallowly, since now the only purpose of unwrapping is to avoid OpenT directly within OpenT.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision:
D59564008

------------------------------------------------------------------------
(from 5131f6f544bfcdaa203cb6ac91745acbe958ba9c)

fbshipit-source-id: 975bba9a51daae818a78c60a5cf90a2abf991319