[flow][refactor] Make reason no longer required for SpeculationKit.(try_custom|try_singleton_throw_on_failure)
Summary:
It's only used for the union and intersection case, so we only need to require it for these two cases and encode them in `UnionCases` and `IntersectionCases`. This allows us to simplify the API for the other case so we don't need to find a reason that's not even used.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D59652432

fbshipit-source-id: 9d10e21572bf6fb26c1cdd5af204e722ff8f5ce1