[flow][refactor] Generalize speculation to allow arbitrary intersection-like cases 2/2: Introduce `CustomCase` and `try_custom`
Summary:
Currently, speculation can only choose between a list of flows. This restriction is no longer fundamental, since we have LTI and much better fully-resolved-ness guarantees. I want to make it support arbitrary intersection-like cases, and use the speculation mechanism to choose one of them based on presence/absence of errors.

This diff adds the support for the general case. With the refactors done previously, we are not really adding any major new capabilities. All we have to is to generalize the flow action with an arbitrary function.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59652426

fbshipit-source-id: f33bdf0e5cf570b2d79cf766003db6cb20a0db93