[flow][refactor] Generalize speculation to allow arbitrary intersection-like cases 1/2: introduce `FlowCase`
Summary:
Currently, speculation can only choose between a list of flows. This restriction is no longer fundamental, since we have LTI and much better fully-resolved-ness guarantees. I want to make it support arbitrary intersection-like cases, and use the speculation mechanism to choose one of them based on presence/absence of errors.

This diff is a prep step to make the next diff easier to review: I first introduced `FlowCase` which encodes the idea that the job we are speculatively running is a flow. In the next diff, it will be extended to support arbitrary functions.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59652428

fbshipit-source-id: adb26445f85438e40b5004f064e81648527efeef