[flow] Replace ArithT, ComparatorT, UnaryArithT with operations on fully resolved types
Summary:
This diff moves ArithT, ComparatorT, UnaryArithT out of the constraint system and we check the type in fully resolved and concretized form.

This sounds like an easy task: we can use the concretization primitive we added in the previous diff, and then run the existing checking function in flow_js_utils on concretized type. However, there is one catch: we need to be very careful with intersection types.

Consider

```
// the weird type below do appear: it's usually a result of refinement with type guard
declare const foo: (?"") & string;
foo + ''; // ok
```

The concretization needs to preserve intersection, because otherwise we will just see `'' | null | void` for LHS, which will fail the type check. Previous diff already ensures that. This diff first adds a distribute union and intersection util in `type_operation_utils`, so that we can preserve the existing intersection behavior that only one branch needs to pass. Then we use this utility to run the check.

The rest of the code is very straightforward: deleting no longer used constraints.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59607509

fbshipit-source-id: 715c548ae6dcaf67d418fa6456b2e129227f716f