[flow] Replace ImplicitVoidReturnT with concretization + empty result test
Summary:
ImplicitVoidT is doing very simple things that no longer need to be done in constraint system: some types that are either void or empty will be flow to it. If empty is flow to it, it will do nothing due to empty shortcut. Else, it will do some simple action. We can replace it with just concretization.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59664912

fbshipit-source-id: 757a98ae1227cd0b27dcb1cf07ac6f4ccbbe6453