[flow] move template coercion check out of constraint system
Summary:
There is a subtyping rule in Subtyping_kit that relies on the use_op of the RHS to bypass a NumT ~> StrT check when this is an argument to a template string.

This diff moves the check to statement.ml by inspecting the type of the argument expression and skipping the case where it is a number.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D59707709

fbshipit-source-id: 5770c893c89dde1f1ae9d9e33f9b3a007f664776