[flow][strutil] Add second (optional) type argument to `StringPrefix` utility type: type of the remainder of the string
Summary:
Changelog: [feature] You can now pass in an (optional) second type argument to the `StringPrefix` utility type. This is the type of the remainder of the string after the prefix is removed. For example, `"$1"`, `"$2"` are subtypes of `StringPrefix<'$', '1' | '2'>`, but `"$999"` is not. `StringPrefix<X>` is equivalent to `StringPrefix<X, string>`

After we add a `StringSuffix` type, this will allow to combine both types and match against both a prefix and suffix on one string.

Reviewed By: panagosg7

Differential Revision: D59615606

fbshipit-source-id: bb01c91954c260c8fddbc3980f303afd1055a5cf