[flow][str] Add `StringSuffix` type which represents strings with a specified suffix
Summary: Changelog: [feature] Added a `StringSuffix` type, which goes along with the recently added `StringPrefix` type. `StringSuffix` can be used to type strings with a specified suffix. E.g. `StringSuffix<'!'>` allows for `'yay!'` and `'woo!'`. The type argument for the suffix must be a string literal. The second, optional, type argument can be used for the type of the remainder of the string after the suffix is removed.

Reviewed By: SamChou19815

Differential Revision: D59817441

fbshipit-source-id: 474eccd127a77bfb17f43d5cd1bceb63c6e67a09