[flow] Upsert tvar in annotation inference
Summary:
While working in some of the later diffs, I have observed some bad interactions in two of the `Context.add_tvar` in annotation_inference.

Tvars in annotation inference starts in unresolved state, and got resolved by resolve_id in a single step. When we create the tvar, we of course add it to the tvar graph; when it's the time to resolve it, we override the constraint in the tvar graph with the resolved variant.

These kinds of writes do look sketchy to me, but I think it's actually fine, as long as in between the writes, the tvar is never forced and inspected. However, this is not true, especially given the fact that we will soon cyclic tvar detection that will force the tvars. In the comment I added, I explained how it can mess up some cases.

We can avoid the issue by only doing the second write if the tvar is not forced. If the tvar is forced at this point, it's certainly the case that the tvar is cyclic and has already become any, so we should keep it in this state, rather than overriding it with something that might bring back the bad cycle.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D59839856

fbshipit-source-id: f707bcc6c1bcf2137762c5fd07913d3dabadeeb2