[flow] Detect all bad cyclic types (almost)
Summary:
In this diff, we start to reuse the defense mechanism first created in D59564008, to make it apply to type sig as well.

The difference is very minimal: for the type sig case, we use a different function to add cyclic type errors, and we can do it for all tvars.

With this, we almost solved the problem of cyclic types. I said almost because the detection problem is still not solved for TypeAppT, but since TypeAppT has its own defense against cyclic TypeAppT, we should have perfect defense when we are operating on bad cyclic types (modulo bugs).

Changelog: [errors] Flow can now detect invalid cyclic types across module boundaries. e.g.

```
// b.js
import type {A} from './a';
export type B = A; // error

// a.js
import type {B} from './b';
export type A = B;
```

Reviewed By: mvitousek

Differential Revision: D59697870

fbshipit-source-id: c5747f1ab9ff0d4e82cd78b0f73f6ab52bf26a7f