[flow] Add Missing Attributes quickfix
Summary:
Adds an "Add Missing Attributes" quickfix for JSX elements.

In code like
```
declare component Foo(foo: number, bar?: string);
const x = <Foo />;
//         ^^^
//       quickfix available here
```
a quickfix is available over the name of the JSX element. If selected it will add `foo={foo}` in the opening header of the element.

The quickfix adds only require properties that are missing.

Implementation: we get the type of the component off the typed AST, and call `mk_type_destructor ReactElementConfigType` to get the config type. We expect this type to be an `ObjT` after the type has been concretized and simplified (`possible_concrete_types_for_operators_checking`). From there we extract all non-optional properties.

We compute the existing attributes. For example in `<Foo bar={bar} {...{ baz: 1 }} />`, the attributes `bar` and `baz` are pre-existing. These are excluded from the solution set.

The quickfix is available by default, but I've added an `add_missing_attributes_quickfix` flowconfig option to disable if necessary during rollout.

Changelog: [feature] An "Add Missing Attributes" quickfix for JSX elements is added. This is available when the cursor is over the name of a JSX element that is missing a required attribute.

Reviewed By: SamChou19815

Differential Revision: D59716473

fbshipit-source-id: ea2682a732e8d47826916107e68bdc5f3ec1a622