[flow][glean] Add `--glean-log` option to `flow glean`
Summary:
To help debug issues with running the Glean runner, add a `--glean-log` option to `flow glean`.
When enabled, this will print the file currently being analyzed, and also information about each type normalization.
Initially I called this `--verbose`, but this also enabled `--verbose` for Flow itself, which was not the intention.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D59890302

fbshipit-source-id: 43883cc30013810228a75611f56bcfe709819a35