[flow][libdef] Special case Object.assign only in statement.ml, instead of exposing the custom function as a type
Summary:
This diff aligns the special casing behavior with other object functions, by only special casing it in statement.ml, and give a less precise and accurate typing for it if you try to use it in other weird ways.

The ultimate motivation is to move all of the constraints (e.g. `ObjectAssignToT`, `ObjectAssignFromT`) that's used to support this special casing out of the constraint system.

Changelog: [libdef] `Object.assign` is still special cased, but the special-casing is only available for a syntactic `Object.assign(...)` call. If you try to use Object.assign in some other ways (e.g. `const f = Object.assign; f(...)`), you will get a a less precise and accurate typing for it. If you don't like the default fallback type, you can override the `Object$Assign` type. While the special-casing behavior will stay for the syntactic call, you should migrate your code to use the spread syntax instead.

Reviewed By: gkz

Differential Revision: D59896190

fbshipit-source-id: d1296b1c7559e78c6d7fae68c066dd1541ab9932