[flow] Use `FilterOptionalT` instead of `PredicateT (NotP VoidP)` for default destructuring
Summary:
I'm trying to move PredicateT out of constraint system, but first I need to unentangle all the complex interactions it has with flow_js.

After the previous diff, PredicateT is only constructed in flow_js in very few places, and this is one of them. It's not necessary since we can use `FilterOptionalT` instead.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D59923462

fbshipit-source-id: d90b1c900e1c0f4a2d03eebd694992939f3c7e9d