[flow] Un-CustomFunction `$Flow$DebugSleep` and `$FlowDebugThrow`
Summary:
I see a path to kill all of the custom functions, so we might as well start to deal with some easy ones.

For the ones I changed in this diff, we can just special case it in statement.ml for our debugging purposes. I just make the name more scary to prevent usages of these types in production.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D59936418

fbshipit-source-id: 955c2e8f0aea86d1a74ff4bf254f6bf7eae19452