[flow][refactor] Kill NoP
Summary:
It turns out that we don't need this at all, especially the special handling in the constraint system, since we can make decision locally in statement.ml and type_env. We just need to make the places that used to return predicate that might contain `NoP` to return an optional predicate instead.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D59932724

fbshipit-source-id: efc7cc8da58845deef70e347ab7d68c8eba0e528