[flow][refactor] Move logic of handling CallLatentPredT into predicate_kit
Summary:
Following D59891711, we also move the logic for CallLatentPredT into predicate_kit. With this change, all the constructions for PredicateT and CallLarentPredT have been moved into predicate_kit, and it's now clear that we are only using flow_js as a concretization engine for refinement. Starting from the next diff, we are going to start to take advantage of that.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59944467

fbshipit-source-id: d5fc96f93866fe5507460beb193afd286abf31b3